fix(dtype): np.min_scalar_type — stop avoiding int8/float16 (NumPy parity)
Investigating the audit item "np.array(byte[]) -> int8": that path is already
correct (C# byte -> uint8/Byte, sbyte -> int8/SByte across every np.array overload,
and AsNumpyDtypeName maps them right). The real, closely-related defect is in
np.min_scalar_type, which falsely assumed "NumSharp doesn't have int8/sbyte" and
never returned float16 (Half) — both types that DO exist among the 15 dtypes.

Bugs fixed (verified against NumPy 2.4.2 convert_datatype.c:min_scalar_type_num):

1. Negative int8-range scalars widened to Int16 instead of SByte (int8).
   min_scalar_type(-1..-128) returned Int16; NumPy returns int8. Affected every
   signed input (sbyte/short/int/long) since they route through MinTypeForSignedInt.
   The sbyte arm (sb>=0 ? Byte : Int16) had the same widening.

2. Float/double values NEVER returned Half (float16). The old code used
   exact-float32-representability, which both (a) missed float16 entirely and
   (b) diverged from NumPy, e.g. min_scalar_type(0.1) returned Double. NumPy demotes
   floats by RANGE (magnitude), allowing precision loss:
     |v| < 65000 (or non-finite) -> float16 ; |v| < 3.4e38 -> float32 ; else float64.
   Bounds are exclusive and copied verbatim from NumPy (65000, not 65504).

3. Added explicit Half input -> Half (NumPy NPY_HALF: float16 stays float16).

promote_types/result_type were checked and are NOT affected (they promote existing
types, not value-based smallest-type inference) — 11/11 narrow-type pairs match NumPy.

Validation:
- New NumPy-oracle differential over 122 cases (all 15-dtype boundaries incl. int8
  negatives, float16 range, exclusive cutoffs) — 0 divergences.
- Corrected two tests that pinned the old buggy behavior (min_scalar_type(-1)->Int16,
  1.0f->Single, NaN/Inf->Single, -10->Int16) to NumPy truth, per project DOD.
- Added np.array C#-type -> NumPy-dtype-name regression (byte[] -> uint8 never int8,
  sbyte[] -> int8, full 12-type matrix) to lock the originally-reported item.
- Full suite green: 11259 passed / 0 failed (net10.0); affected classes green on net8.0.

+++ test/NumSharp.UnitTest/APIs/np.typing.Test.cs
@@ -292,8 +292,8 @@ public class NpTypingTests
     [TestMethod]
     public void MinScalarType_SmallNegative()
     {
-        // No Int8 in NumSharp, so smallest signed is Int16
-        np.min_scalar_type(-10).Should().Be(NPTypeCode.Int16);
+        // NumPy 2.4.2: small negatives demote to int8 (SByte). NumSharp HAS int8.
+        np.min_scalar_type(-10).Should().Be(NPTypeCode.SByte);
     }
 
     [TestMethod]
